blackacre/requirements.txt
grpcio-tools==1.33.2
spacy==2.3.4